Fix AIR SDK compiler bug
This fixes an issue that was causing compc to produce a
corrupted swc, that would then cause adt to crash during the
iOS packaging process.

The AIR compiler bug is related to switch statements - empty
switch statements cause problems, and in this case a nested
switch without a default clause seems to be the culprit.

Two threads on the Adobe forums that reference this bug:
https://forums.adobe.com/thread/1641400
https://forums.adobe.com/message/7110611